Pull blame setting code into base class.
