+++ src/main/java/hudson/plugins/analysis/util/AbstractBlamer.java
@@ -6,10 +6,15 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+import org.apache.commons.lang3.exception.ExceptionUtils;
+
+import jenkins.MasterToSlaveFileCallable;
+
 import hudson.EnvVars;
 import hudson.FilePath;
 import hudson.model.TaskListener;
 import hudson.plugins.analysis.util.model.FileAnnotation;
+import hudson.remoting.VirtualChannel;
 
 /**
  * A base class for {@link Blamer} implementations.
@@ -33,6 +38,64 @@ public abstract class AbstractBlamer implements Blamer {
         this.workspace = workspace;
         this.listener = listener;
     }
+
+    @Override
+    public void blame(final Set<FileAnnotation> annotations) {
+        try {
+            if (annotations.isEmpty()) {
+                return;
+            }
+
+            invokeBlamer(annotations);
+        }
+        catch (IOException e) {
+            error("Computing blame information failed with an exception:%n%s%n%s",
+                    e.getMessage(), ExceptionUtils.getStackTrace(e));
+        }
+        catch (InterruptedException e) {
+            // nothing to do, already logged
+        }
+    }
+
+    private void invokeBlamer(final Set<FileAnnotation> annotations) throws IOException, InterruptedException {
+        final Map<String, BlameRequest> linesOfConflictingFiles = extractConflictingFiles(annotations);
+
+        Map<String, BlameRequest> blamesOfConflictingFiles = getWorkspace().act(new MasterToSlaveFileCallable<Map<String, BlameRequest>>() {
+            @Override
+            public Map<String, BlameRequest> invoke(final File workspace, final VirtualChannel channel) throws IOException, InterruptedException {
+                return blame(linesOfConflictingFiles);
+            }
+        });
+
+        setBlameResults(annotations, blamesOfConflictingFiles);
+    }
+
+    private void setBlameResults(final Set<FileAnnotation> annotations, final Map<String, BlameRequest> blamesOfConflictingFiles) {
+        for (FileAnnotation annotation : annotations) {
+            if (blamesOfConflictingFiles.containsKey(annotation.getFileName())) {
+                BlameRequest blame = blamesOfConflictingFiles.get(annotation.getFileName());
+                int line = annotation.getPrimaryLineNumber();
+                annotation.setAuthorName(blame.getName(line));
+                annotation.setAuthorEmail(blame.getEmail(line));
+                annotation.setCommitId(blame.getCommit(line));
+            }
+            else {
+                log("Skipping file %s, no result found.%n", annotation.getFileName());
+            }
+        }
+    }
+
+    /**
+     * Computes for each conflicting file a {@link BlameRequest}. Note that this call is executed on a build agent.
+     * I.e., the blamer instance and all transfer objects need to be serializable.
+     *
+     * @param linesOfConflictingFiles a mapping of file names to blame request. Each blame request defines the lines
+     *                                that need to be mapped to an author.
+     * @return the same mapping, now filled with blame information
+     * @throws InterruptedException if this operation has been canceled
+     * @throws IOException          in case of an error
+     */
+    protected abstract Map<String, BlameRequest> blame(Map<String, BlameRequest> linesOfConflictingFiles) throws InterruptedException, IOException;
 
     /**
      * Extracts the relative file names of the files that contain annotations
@@ -106,14 +169,29 @@ public abstract class AbstractBlamer implements Blamer {
         listener.error("<Git Blamer> " + String.format(message, args));
     }
 
+    /**
+     * Returns the workspace path on the agent.
+     *
+     * @return workspace path
+     */
     protected FilePath getWorkspace() {
         return workspace;
     }
 
+    /**
+     * Returns the listener for logging statements.
+     *
+     * @return logger
+     */
     protected TaskListener getListener() {
         return listener;
     }
 
+    /**
+     * Returns the environment on the agent.
+     *
+     * @return the environment
+     */
     protected EnvVars getEnvironment() {
         return environment;
     }
